# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'LearnHub API Documentation'
  description: 'The documentation'
  version: 1.0.0
servers:
  -
    url: 'https://staging-api.learnhub.mk'
paths:
...
  /content/blog-posts:
    get:
      summary: 'Blog posts'
      operationId: blogPosts
      description: 'This endpoint list all blog post'
      parameters:
        -
          in: query
          name: title
          description: ''
          example: learnhub
          required: false
          schema:
            type: string
            description: ''
            example: learnhub
            nullable: false
        -
          in: query
          name: tags
          description: ''
          example: '[php,laravel,react]'
          required: false
          schema:
            type: string
            description: ''
            example: '[php,laravel,react]'
            nullable: false
        -
          in: query
          name: author
          description: ''
          example: john
          required: false
          schema:
            type: string
            description: ''
            example: john
            nullable: false
        -
          in: query
          name: sort
          description: ''
          example: title
          required: false
          schema:
            type: string
            description: ''
            example: title
            nullable: false
            enum:
              - id
              - title
              - publish_date
              - created_at
        -
          in: query
          name: per_page
          description: ''
          example: 3
          required: false
          schema:
            type: integer
            description: ''
            example: 3
            nullable: false
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: ''
                  example: archived
                  nullable: false
                  enum:
                    - draft
                    - published
                    - in_review
                    - archived
                publish_date:
                  type: string
                  description: 'Must be a valid date in the format <code>Y-m-d H:i:s</code>.'
                  example: '2024-12-01 02:38:26'
                  nullable: true
              required:
                - status
    post:
      summary: 'Create Blog posts'
      operationId: createBlogPosts
      description: 'This endpoint will create a single blog post'
      parameters: []
      responses: {  }
      tags:
        - Content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: ''
                  example: 'Example Blog Post Title '
                  nullable: false
                excerpt:
                  type: string
                  description: ''
                  example: 'This is test blogpost example'
                  nullable: false
                content:
                  type: string
                  description: ''
                  example: 'Lorem ipsum dolor sit amet, consectetur adipiscing ...'
                  nullable: false
                tags:
                  type: array
                  description: ''
                  example: '[1,2,3]'
                  items:
                    type: string
              required:
                - title
                - excerpt
                - content
                - tags
  '/content/blog-posts/{id}':
    get:
      summary: 'Blog post'
      operationId: blogPost
      description: 'This endpoint returns a single blog post'
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: ''
                  example: published
                  nullable: false
                  enum:
                    - draft
                    - published
                    - in_review
                    - archived
                publish_date:
                  type: string
                  description: 'Must be a valid date in the format <code>Y-m-d H:i:s</code>.'
                  example: '2024-12-01 02:38:27'
                  nullable: true
              required:
                - status
    put:
      summary: 'Update Blog post'
      operationId: updateBlogPost
      description: 'This endpoint will update a single blog post'
      parameters: []
      responses: {  }
      tags:
        - Content
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: ''
                  example: 'Example Blog Post Title '
                  nullable: false
                excerpt:
                  type: string
                  description: ''
                  example: 'This is test blogpost example'
                  nullable: false
                slug:
                  type: string
                  description: ''
                  example: 'example-blog-post-title '
                  nullable: false
                content:
                  type: string
                  description: ''
                  example: 'Lorem ipsum dolor sit amet, consectetur adipiscing ...'
                  nullable: false
                tags:
                  type: array
                  description: ''
                  example: '[1,2,3]'
                  items:
                    type: string
    delete:
      summary: 'Delete Blog posts'
      operationId: deleteBlogPosts
      description: 'This endpoint deletes blog post'
      parameters: []
      responses: {  }
      tags:
        - Content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  description: ''
                  example: archived
                  nullable: false
                  enum:
                    - draft
                    - published
                    - in_review
                    - archived
                publish_date:
                  type: string
                  description: 'Must be a valid date in the format <code>Y-m-d H:i:s</code>.'
                  example: '2024-12-01 02:38:27'
                  nullable: true
              required:
                - status
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the blog post.'
        example: non
        required: true
        schema:
          type: string
...
  /blog-posts:
    get:
      summary: 'Blog posts'
      operationId: blogPosts
      description: 'This endpoint list all blog post from newest ones to the oldest.Additionally you may use ?title= query param to filter by title'
      parameters:
        -
          in: query
          name: title
          description: ''
          example: sapiente
          required: false
          schema:
            type: string
            description: ''
            example: sapiente
            nullable: false
        -
          in: query
          name: per_page
          description: ''
          example: 14
          required: false
          schema:
            type: integer
            description: ''
            example: 14
            nullable: false
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: []
                  links:
                    first: 'http://learnhub.test/blog-posts?page=1'
                    last: 'http://learnhub.test/blog-posts?page=1'
                    prev: null
                    next: null
                  meta:
                    current_page: 1
                    from: null
                    last_page: 1
                    links:
                      -
                        url: null
                        label: '&laquo; Претходна'
                        active: false
                      -
                        url: 'http://learnhub.test/blog-posts?page=1'
                        label: '1'
                        active: true
                      -
                        url: null
                        label: 'Следна &raquo;'
                        active: false
                    path: 'http://learnhub.test/blog-posts'
                    per_page: 14
                    to: null
                    total: 0
                properties:
                  data:
                    type: array
                    example: []
                  links:
                    type: object
                    properties:
                      first:
                        type: string
                        example: 'http://learnhub.test/blog-posts?page=1'
                      last:
                        type: string
                        example: 'http://learnhub.test/blog-posts?page=1'
                      prev:
                        type: string
                        example: null
                      next:
                        type: string
                        example: null
                  meta:
                    type: object
                    properties:
                      current_page:
                        type: integer
                        example: 1
                      from:
                        type: string
                        example: null
                      last_page:
                        type: integer
                        example: 1
                      links:
                        type: array
                        example:
                          -
                            url: null
                            label: '&laquo; Претходна'
                            active: false
                          -
                            url: 'http://learnhub.test/blog-posts?page=1'
                            label: '1'
                            active: true
                          -
                            url: null
                            label: 'Следна &raquo;'
                            active: false
                        items:
                          type: object
                          properties:
                            url:
                              type: string
                              example: null
                            label:
                              type: string
                              example: '&laquo; Претходна'
                            active:
                              type: boolean
                              example: false
                      path:
                        type: string
                        example: 'http://learnhub.test/blog-posts'
                      per_page:
                        type: integer
                        example: 14
                      to:
                        type: string
                        example: null
                      total:
                        type: integer
                        example: 0
      tags:
        - Website
      security: []
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []